Refactor: Moved files touched by the Tasks-navigation PR (NP-51254)
