+++ src/components/ticket-list/TicketList.tsx
@@ -5,29 +5,29 @@ import { useEffect, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSelector } from 'react-redux';
 import { useLocation, useNavigate } from 'react-router';
-import { SortOrder, TicketOrderBy, TicketSearchParam } from '../../../api/searchApi';
-import { AreaOfResponsibilitySelector } from '../../../components/AreaOfResponsibiltySelector';
-import { CategorySearchFilter } from '../../../components/CategorySearchFilter';
-import { CuratorSelector } from '../../../components/CuratorSelector';
-import { DialoguesWithoutCuratorButton } from '../../../components/DialoguesWithoutCuratorButton';
-import { ErrorBoundary } from '../../../components/ErrorBoundary';
-import { HeadTitle } from '../../../components/HeadTitle';
-import { ListPagination } from '../../../components/ListPagination';
-import { ListSkeleton } from '../../../components/ListSkeleton';
-import { SearchForm } from '../../../components/SearchForm';
-import { SortSelector } from '../../../components/_molecules/SortSelector';
-import { TicketStatusFilter } from '../../../components/TicketStatusFilter';
-import { RootState } from '../../../redux/store';
-import { CustomerTicketSearchResponse, ticketStatusValues } from '../../../types/publication_types/ticket.types';
-import { RoleName } from '../../../types/user.types';
-import { ROWS_PER_PAGE_OPTIONS } from '../../../utils/constants';
-import { dataTestId } from '../../../utils/dataTestIds';
-import { stringIncludesMathJax, typesetMathJax } from '../../../utils/mathJaxHelpers';
-import { syncParamsWithSearchFields } from '../../../utils/searchHelpers';
-import { UrlPathTemplate } from '../../../utils/urlPaths';
-import { ExcludeSubunitsCheckbox } from './ExcludeSubunitsCheckbox';
-import { TicketDateIntervalFilter } from './TicketDateIntervalFilter';
-import { TicketListItem } from './TicketListItem';
+import { SortOrder, TicketOrderBy, TicketSearchParam } from '../../api/searchApi';
+import { AreaOfResponsibilitySelector } from '../AreaOfResponsibiltySelector';
+import { CategorySearchFilter } from '../CategorySearchFilter';
+import { CuratorSelector } from '../CuratorSelector';
+import { DialoguesWithoutCuratorButton } from '../DialoguesWithoutCuratorButton';
+import { ErrorBoundary } from '../ErrorBoundary';
+import { HeadTitle } from '../HeadTitle';
+import { ListPagination } from '../ListPagination';
+import { ListSkeleton } from '../ListSkeleton';
+import { SearchForm } from '../SearchForm';
+import { SortSelector } from '../_molecules/SortSelector';
+import { TicketStatusFilter } from '../TicketStatusFilter';
+import { RootState } from '../../redux/store';
+import { CustomerTicketSearchResponse, ticketStatusValues } from '../../types/publication_types/ticket.types';
+import { RoleName } from '../../types/user.types';
+import { ROWS_PER_PAGE_OPTIONS } from '../../utils/constants';
+import { dataTestId } from '../../utils/dataTestIds';
+import { stringIncludesMathJax, typesetMathJax } from '../../utils/mathJaxHelpers';
+import { syncParamsWithSearchFields } from '../../utils/searchHelpers';
+import { UrlPathTemplate } from '../../utils/urlPaths';
+import { ExcludeSubunitsCheckbox } from '../../pages/messages/components/ExcludeSubunitsCheckbox';
+import { TicketDateIntervalFilter } from '../../pages/messages/components/TicketDateIntervalFilter';
+import { TicketListItem } from './_components/TicketListItem';
 
 interface TicketListProps {
   ticketsQuery: UseQueryResult<CustomerTicketSearchResponse>;
