+++ src/pages/tasks/TasksPage.tsx
@@ -27,11 +27,11 @@ import { NviCorrectionList } from '../messages/components/NviCorrectionList';
 import { NviCorrectionListNavigationAccordion } from '../messages/components/NviCorrectionListNavigationAccordion';
 import { NviDisputePage } from '../messages/components/NviDisputePage';
 import { ResultRegistrationsNavigationListAccordion } from '../messages/components/ResultRegistrationsNavigationListAccordion';
-import { TicketList } from '../messages/components/TicketList';
+import { TicketList } from '../../components/ticket-list/TicketList';
 import { TicketTypeTag } from '../messages/components/TicketTypeTag';
 import { checkPages } from '../messages/tasks-helpers';
 import { useGetNotificationCounts, useGetTicketsCounts } from '../messages/user-dialog-helpers';
-import { RegistrationLandingPage } from '../public_registration/RegistrationLandingPage';
+import { RegistrationLandingPage } from '../../components/registration-landing-page/RegistrationLandingPage';
 import { NviCandidatesNavigationAccordion } from './_components/NviCandidatesNavigationAccordion';
 import { NviCandidatePage } from './nvi/nvi-candidate-page/NviCandidatePage';
 import { NviCandidatesListPage } from './nvi/NviCandidatesListPage';
